API reference

API reference

Input widgets

InteractBase.input — Function.

input(o; typ="text")

Create an HTML5 input element of type type (e.g. "text", "color", "number", "date") with o as initial value.

source

InteractBase.button — Function.

button(content=""; clicks::Observable)

A button. content goes inside the button. Note the button content supports a special clicks variable, e.g.: button("clicked {{clicks}} times")

source

InteractBase.filepicker — Function.

filepicker(label=""; placeholder="", multiple=false, accept="*")

Create a widget to select files. If multiple=true the observable will hold an array containing the paths of all selected files. Use accept to only accept some formats, e.g. accept=".csv"

source

InteractBase.textbox — Function.

textbox(label=""; text::Union{String, Observable})

Create a text input area with an optional label e.g. textbox("enter number:")

source

InteractBase.autocomplete — Function.

autocomplete(options, o=""; multiple=false, accept="*")

Create a textbox input with autocomplete options specified by options and with o as initial value.

source

InteractBase.checkbox — Function.

checkbox(checked::Union{Bool, Observable}=false; label)

A checkbox. e.g. checkbox(label="be my friend?")

source

InteractBase.toggle — Function.

toggle(checked::Union{Bool, Observable}=false; label)

A toggle switch. e.g. toggle(label="be my friend?")

source

InteractBase.slider — Function.

function slider(vals; # Range
                value=medianelement(valse),
                label="", kwargs...)

Creates a slider widget which can take on the values in vals, and updates observable value when the slider is changed:

source

Option widgets

InteractBase.dropdown — Function.

dropdown(options::Associative;
         value = first(values(options)),
         label = nothing,
         multiple = false)

A dropdown menu whose item labels will be the keys of options. If multiple=true the observable will hold an array containing the values of all selected items e.g. dropdown(OrderedDict("good"=>1, "better"=>2, "amazing"=>9001))

source

InteractBase.togglebuttons — Function.

togglebuttons(options::Associative; selected::Union{T, Observable})

Creates a set of toggle buttons whose labels will be the keys of options.

source

InteractBase.radiobuttons — Function.

radiobuttons(options::Associative;
             value::Union{T, Observable} = first(values(options)))

e.g. radiobuttons(OrderedDict("good"=>1, "better"=>2, "amazing"=>9001))

source