API reference

API reference

Text input

These are widgets to select text input that's typed in by the user. For numbers use spinbox and for strings use textbox.

InteractBase.spinbox — Function.

spinbox(label=""; value=nothing) Create a widget to select numbers with placeholder label

source

InteractBase.textbox — Function.

textbox(label=""; text::Union{String, Observable})

Create a text input area with an optional label e.g. textbox("enter number:"). Use typ=... to specify the type of text. For example typ="email" or typ=password

source

InteractBase.autocomplete — Function.

autocomplete(options, label=""; value="")

Create a textbox input with autocomplete options specified by options, with value as initial value and label as label.

source

Type input

These are widgets to select a specific, non-text, type of input. So far, Date, Time, Color and Bool are supported.

InteractBase.datepicker — Function.

datepicker(; value=nothing)

Create a widget to select dates.

source

InteractBase.timepicker — Function.

timepicker(; value=nothing)

Create a widget to select times.

source

InteractBase.colorpicker — Function.

colorpicker(; value=nothing)

Create a widget to select colors.

source

InteractBase.checkbox — Function.

checkbox(checked::Union{Bool, Observable}=false; label)

A checkbox. e.g. checkbox(label="be my friend?")

source

InteractBase.toggle — Function.

toggle(checked::Union{Bool, Observable}=false; label)

A toggle switch. e.g. toggle(label="be my friend?")

source

File input

InteractBase.filepicker — Function.

filepicker(label="Choose a file..."; multiple=false, accept="*")

Create a widget to select files. If multiple=true the observable will hold an array containing the paths of all selected files. Use accept to only accept some formats, e.g. accept=".csv"

source

Range input

InteractBase.slider — Function.

function slider(vals::Range; # Range
                value=medianelement(valse),
                label="", kwargs...)

Creates a slider widget which can take on the values in vals, and updates observable value when the slider is changed:

source

Callback input

InteractBase.button — Function.

button(label=""; clicks::Observable)

A button. content goes inside the button. Note the button content supports a special clicks variable, e.g.: button("clicked {{clicks}} times")

source

HTML5 input

All of the inputs above are implemented wrapping the input tag of HTML5 which can be accessed more directly as follows:

InteractBase.input — Function.

input(o; typ="text")

Create an HTML5 input element of type type (e.g. "text", "color", "number", "date") with o as initial value.

source

Option widgets

InteractBase.dropdown — Function.

dropdown(options::Associative;
         value = first(values(options)),
         label = nothing,
         multiple = false)

A dropdown menu whose item labels will be the keys of options. If multiple=true the observable will hold an array containing the values of all selected items e.g. dropdown(OrderedDict("good"=>1, "better"=>2, "amazing"=>9001))

source

dropdown(values::AbstractArray; kwargs...)

dropdown with labels string.(values) see dropdown(options::Associative; ...) for more details

source

InteractBase.togglebuttons — Function.

togglebuttons(options::Associative; value::Union{T, Observable})

Creates a set of toggle buttons whose labels will be the keys of options.

source

togglebuttons(values::AbstractArray; kwargs...)

togglebuttons with labels string.(values) see togglebuttons(options::Associative; ...) for more details

source

InteractBase.radiobuttons — Function.

radiobuttons(options::Associative;
             value::Union{T, Observable} = first(values(options)))

e.g. radiobuttons(OrderedDict("good"=>1, "better"=>2, "amazing"=>9001))

source

radiobuttons(values::AbstractArray; kwargs...)

radiobuttons with labels string.(values) see radiobuttons(options::Associative; ...) for more details

source